refactor(tigrbl_auth): update rfc module imports
